Refactor component imports for better organization and clarity
