+++ apps/nextjs/src/app/(dashboard)/dashboard/layout.tsx
@@ -8,7 +8,7 @@ import {
 import type { ChildrenProps } from "~/types";
 import DashboardContentSkeleton from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-skeleton";
 import DashboardHeader from "~/app/(dashboard)/dashboard/_components/content/dashboard-header";
-import { BreadCrumbsProvider } from "./_components/breadcrumbs/context";
+import { BreadCrumbsProvider } from "~/components/breadcrumbs/context";
 
 interface DashboardLayoutProps extends ChildrenProps {
   sidebar: React.ReactNode;
