+++ apps/nextjs/src/app/layout.tsx
@@ -15,7 +15,7 @@ import { GeistSans } from "geist/font/sans";
 import { cn, ThemeProvider } from "@battle-stadium/ui";
 
 import type { ChildrenProps } from "~/types";
-import { AdSenseScript } from "~/app/_components/ad-sense";
+import { AdSenseScript } from "~/components/google-adsense/ad-sense";
 import { env } from "~/env";
 import { siteConfig } from "~/lib/config/site";
 import { TRPCReactProvider } from "~/trpc/server";
