+++ apps/nextjs/src/components/google-adsense/ad-banner.tsx
