+++ apps/nextjs/src/components/google-adsense/ad-sense.tsx
